+++ src/index.js
@@ -73,6 +73,7 @@ class RelationshipGraph {
             valueKeyName: userConfig.valueKeyName // Set a custom key in the tooltip.
         };
 
+        // TODO: Find a better way to handles these.
         if (this.configuration.showTooltips === undefined) {
             this.configuration.showTooltips = true;
         }
@@ -86,7 +87,7 @@ class RelationshipGraph {
         }
 
         // If the threshold array is made up of numbers, make sure that it is sorted.
-        if (this.configuration.thresholds.length > 0 && (typeof this.configuration.thresholds[0]) == 'number') {
+        if (this.configuration.thresholds.length && (typeof this.configuration.thresholds[0]) == 'number') {
             this.configuration.thresholds.sort();
         }
 
@@ -115,6 +116,14 @@ class RelationshipGraph {
          * @type {number}
          */
         this._spacing = 1;
+
+        /**
+         * Used to determine whether or not d3 V3 or V4 is being used.
+         *
+         * @type {boolean}
+         * @private
+         */
+        this._d3V4 = !!this.configuration.selection._groups;
 
         /**
          * Function to create the tooltip.
@@ -267,13 +276,13 @@ class RelationshipGraph {
     static noop() { }
 
     /**
-     * Returns a sorted Array.
+     * Sorts the array of JSON by parent name. This method is case insensitive.
      *
      * @param json {Array} The Array to be sorted.
      */
     static sortJson(json) {
         json.sort(function(child1, child2) {
-            let parent1 = child1.parent.toLowerCase(),
+            const parent1 = child1.parent.toLowerCase(),
                 parent2 = child2.parent.toLowerCase();
 
             return (parent1 > parent2) ? 1 : (parent1 < parent2) ? -1 : 0;
@@ -292,6 +301,10 @@ class RelationshipGraph {
     static stringCompare(value, thresholds) {
         if (typeof value !== 'string') {
             throw 'Cannot make value comparison between a string and a ' + (typeof value) + '.';
+        }
+
+        if (!thresholds || !thresholds.length) {
+            throw 'Cannot find correct threshold because there are no thresholds.';
         }
 
         const thresholdsLength = thresholds.length;
@@ -319,6 +332,10 @@ class RelationshipGraph {
             throw 'Cannot make value comparison between a number and a ' + (typeof value) + '.';
         }
 
+        if (!thresholds || !thresholds.length) {
+            throw 'Cannot find correct threshold because there are no thresholds.';
+        }
+
         const length = thresholds.length;
 
         for (let i = 0; i < length; i++) {
@@ -337,8 +354,8 @@ class RelationshipGraph {
      * @private
      */
     getId() {
-        const parent = this.configuration.selection._groups ? this.configuration.selection._groups[0][0] :
-            this.configuration.selection[0][0];
+        const selection = this.configuration.selection,
+            parent = this._d3V4 ? selection._groups[0][0] : selection[0][0];
 
         return parent.id;
     }
@@ -385,7 +402,8 @@ class RelationshipGraph {
             previousParent = '',
             parentLength = parents.length,
             {configuration} = this,
-            {blockSize} = configuration;
+            {blockSize} = configuration,
+            {selection} = configuration;
 
         for (i = 0; i < parentLength; i++) {
             let current = parents[i] + ' ( ' + parentSizes[parentNames[i]] + ') ';
@@ -397,8 +415,7 @@ class RelationshipGraph {
 
         // Calculate the row and column for each child block.
         let longestWidth = this.getPixelLength(longest),
-            parentDiv = configuration.selection._groups ? configuration.selection._groups[0][0] :
-                configuration.selection[0][0],
+            parentDiv = this._d3V4 ? selection._groups[0][0] : selection[0][0],
             calculatedMaxChildren = (configuration.maxChildCount === 0) ?
                 Math.floor((parentDiv.parentElement.clientWidth - blockSize - longestWidth) / blockSize) :
                 configuration.maxChildCount,
@@ -488,7 +505,7 @@ class RelationshipGraph {
                 }
             };
 
-            element.__id = this.getId() + '-child-node' + element.__row + element.__index;
+            element.__id = this.getId() + '-child-node' + element.__row + '-' + element.__index;
         }
 
         return [
@@ -771,9 +788,7 @@ class RelationshipGraph {
             maxWidth = longestWidth + calculatedMaxChildren * configuration.blockSize;
 
             // Account for the added _spacing.
-            for (i = 0; i < calculatedMaxChildren; i++) {
-                maxWidth += this._spacing * i;
-            }
+            maxWidth += this._spacing * calculatedMaxChildren;
 
             for (i = 0; i < row; i++) {
                 maxHeight += this._spacing * i;